+++ src/Geta.EmailNotification.Amazon/AmazonEmailNotificationClient.cs
@@ -19,7 +19,7 @@ namespace Geta.EmailNotification.Amazon
             _amazonMessageFactory = amazonMessageFactory;
         }
 
-        public EmailNotificationResponse Send(EmailNotificationRequestBase request)
+        public EmailNotificationResponse Send(EmailNotificationRequest request)
         {
             try
             {
@@ -44,7 +44,7 @@ namespace Geta.EmailNotification.Amazon
             }
         }
 
-        public async Task<EmailNotificationResponse> SendAsync(EmailNotificationRequestBase request)
+        public async Task<EmailNotificationResponse> SendAsync(EmailNotificationRequest request)
         {
             try
             {