+++ src/Geta.EmailNotification.Core/EmailNotificationRequestBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.AspNetCore.Mvc.ViewFeatures;
+using Geta.EmailNotification.Common;
 using MimeKit;
 using AttachmentCollection = MimeKit.AttachmentCollection;
 
@@ -266,23 +266,12 @@ namespace Geta.EmailNotification.Core
         /// </summary>
         /// <param name="dictionary">Dictionary of ViewData values.</param>
         /// <returns>Current EmailNotificationRequestBuilder instance.</returns>
-        public EmailNotificationRequestBuilder WithViewData(ViewDataDictionary dictionary)
+        public EmailNotificationRequestBuilder WithViewData(IDictionary<string, object> dictionary)
         {
             foreach (var pair in dictionary)
             {
                 _request.ViewData.Add(pair);
             }
-            return this;
-        }
-
-        /// <summary>
-        /// Sets Model value on ViewData. Model can be used in Razor view to render email.
-        /// </summary>
-        /// <param name="value">View model's value.</param>
-        /// <returns>Current EmailNotificationRequestBuilder instance.</returns>
-        public EmailNotificationRequestBuilder WithViewModel(object value)
-        {
-            _request.ViewData.Model = value;
             return this;
         }
 
@@ -294,6 +283,17 @@ namespace Geta.EmailNotification.Core
         public EmailNotificationRequestBuilder WithAttachment(MimeEntity attachment)
         {
             _request.Attachments.Add(attachment);
+            return this;
+        }
+        
+        /// <summary>
+        /// Sets Model value on ViewData. Model can be used in Razor view to render email.
+        /// </summary>
+        /// <param name="value">View model's value.</param>
+        /// <returns>Current EmailNotificationRequestBuilder instance.</returns>
+        public EmailNotificationRequestBuilder WithViewModel(object value)
+        {
+            _request.Model = value;
             return this;
         }